Add collection_radio_buttons method.
This works similar to the Rails ActionView::Helpers::FormBuilder method of the
same name, but styles the inputs as per the GOVUK standard. Because it's a
re-implementation it also misses out on some of the features of the Rails
version, such as providing a block to build the individual elements manually.